Dodatno dorađena baza. Popravljen getAllKonf metoda te dodatno sređeni ključevi entiteta Mjesto. Sada Mjesto ima dva glavna ključa, ulica i kucBroj

+++ IzvorniKod/backend/src/main/java/opp/domain/MjestoKonferencijaDTO.java
@@ -25,5 +25,5 @@ public class MjestoKonferencijaDTO {
 
     private String ulica;
 
-    private int kucBroj;
+    private Long kucBroj;
 }
